+++ src/main/java/camp/nextstep/edu/kitchenpos/product/dao/ProductDao.java
@@ -1,6 +1,6 @@
-package camp.nextstep.edu.kitchenpos.dao;
+package camp.nextstep.edu.kitchenpos.product.dao;
 
-import camp.nextstep.edu.kitchenpos.model.Product;
+import camp.nextstep.edu.kitchenpos.product.model.Product;
 
 import java.util.List;
 import java.util.Optional;
