+++ src/main/java/camp/nextstep/edu/kitchenpos/product/model/Product.java
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.kitchenpos.model;
+package camp.nextstep.edu.kitchenpos.product.model;
 
 import java.math.BigDecimal;
