+++ src/test/java/camp/nextstep/edu/kitchenpos/menu/bo/MenuBoTest.java
@@ -3,10 +3,10 @@ package camp.nextstep.edu.kitchenpos.menu.bo;
 import camp.nextstep.edu.kitchenpos.menu.dao.MenuDao;
 import camp.nextstep.edu.kitchenpos.menugroup.dao.MenuGroupDao;
 import camp.nextstep.edu.kitchenpos.menu.dao.MenuProductDao;
-import camp.nextstep.edu.kitchenpos.dao.ProductDao;
+import camp.nextstep.edu.kitchenpos.product.dao.ProductDao;
 import camp.nextstep.edu.kitchenpos.menu.model.Menu;
 import camp.nextstep.edu.kitchenpos.menu.model.MenuProduct;
-import camp.nextstep.edu.kitchenpos.model.Product;
+import camp.nextstep.edu.kitchenpos.product.model.Product;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
